Updating Messages Sandbox snippets
