+++ .env-example
@@ -54,3 +54,10 @@ THEME_ID=THEME_ID
 RECORDING_ID=RECORDING_ID
 EXPIRATION_DATE=2033-07-21T13:37:59.810Z
 LOGO_TYPE=favicon
+
+## For Messages Sandbox
+MESSAGES_SANDBOX_WHATSAPP_NUMBER=MESSAGES_SANDBOX_WHATSAPP_NUMBER
+MESSAGES_SANDBOX_ALLOW_LISTED_TO_NUMBER=MESSAGES_SANDBOX_ALLOW_LISTED_TO_NUMBER
+MESSAGES_SANDBOX_FB_ID=MESSAGES_SANDBOX_FB_ID
+MESSAGES_SANDBOX_ALLOW_LISTED_FB_RECIPIENT_ID=MESSAGES_SANDBOX_ALLOW_LISTED_FB_RECIPIENT_ID
+MESSAGES_SANDBOX_VIBER_SERVICE_ID=MESSAGES_SANDBOX_VIBER_SERVICE_ID
