+++ src/main/java/com/vonage/quickstart/messages/sandbox/messenger/SendMessengerText.java
@@ -19,14 +19,14 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.vonage.quickstart.messages.sandbox;
+package com.vonage.quickstart.messages.sandbox.messenger;
 
 import com.vonage.client.VonageClient;
-import com.vonage.client.messages.whatsapp.WhatsappVideoRequest;
+import com.vonage.client.messages.messenger.MessengerTextRequest;
 import static com.vonage.quickstart.Util.configureLogging;
 import static com.vonage.quickstart.Util.envVar;
 
-public class SendWhatsappVideo {
+public class SendMessengerText {
 
 	public static void main(String[] args) throws Exception {
 		configureLogging();
@@ -37,11 +37,11 @@ public class SendWhatsappVideo {
 				.build()
 				.getMessagesClient()
 				.useSandboxEndpoint()
-				.sendMessage(WhatsappVideoRequest.builder()
-					.from(envVar("VONAGE_WHATSAPP_NUMBER"))
-					.to(envVar("TO_NUMBER"))
-					.url("https://file-examples.com/storage/fee788409562ada83b58ed5/2017/04/file_example_MP4_640_3MG.mp4")
-					.build()
+				.sendMessage(MessengerTextRequest.builder()
+						.from(envVar("MESSAGES_SANDBOX_FB_ID"))
+						.to(envVar("MESSAGES_SANDBOX_ALLOW_LISTED_FB_RECIPIENT_ID"))
+						.text("Don't miss out on our latest offers!")
+						.build()
 				).getMessageUuid()
 		);
 	}
