+++ src/main/java/com/vonage/quickstart/messages/sandbox/viber/SendViberText.java
@@ -19,25 +19,31 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.vonage.quickstart.messages.sandbox;
+package com.vonage.quickstart.messages.sandbox.viber;
 
 import com.vonage.client.VonageClient;
-import com.vonage.client.messages.sms.SmsTextRequest;
-import static com.vonage.quickstart.Util.*;
+import com.vonage.client.messages.viber.Category;
+import com.vonage.client.messages.viber.ViberTextRequest;
+import static com.vonage.quickstart.Util.configureLogging;
+import static com.vonage.quickstart.Util.envVar;
 
-public class SendSmsText {
+public class SendViberText {
 
 	public static void main(String[] args) throws Exception {
+		configureLogging();
+
 		System.out.println(VonageClient.builder()
 				.applicationId(envVar("VONAGE_APPLICATION_ID"))
 				.privateKeyPath(envVar("VONAGE_PRIVATE_KEY_PATH"))
 				.build()
 				.getMessagesClient()
-				.sendMessage(SmsTextRequest.builder()
-					.from("Vonage Java SDK")
-					.to(envVar("TO_NUMBER"))
-					.text("Hello, "+System.getenv("NAME"))
-					.build()
+				.useSandboxEndpoint()
+				.sendMessage(ViberTextRequest.builder()
+						.from(envVar("MESSAGES_SANDBOX_VIBER_SERVICE_ID"))
+						.to(envVar("MESSAGES_SANDBOX_ALLOW_LISTED_TO_NUMBER"))
+						.text("Don't miss out on our latest offers!")
+						.category(Category.PROMOTION)
+						.build()
 				).getMessageUuid()
 		);
 	}
