+++ src/main/java/com/vonage/quickstart/messages/sandbox/whatsapp/SendWhatsappSticker.java
@@ -19,15 +19,14 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.vonage.quickstart.messages.sandbox;
+package com.vonage.quickstart.messages.sandbox.whatsapp;
 
 import com.vonage.client.VonageClient;
-import com.vonage.client.messages.viber.Category;
-import com.vonage.client.messages.viber.ViberTextRequest;
+import com.vonage.client.messages.whatsapp.WhatsappStickerRequest;
 import static com.vonage.quickstart.Util.configureLogging;
 import static com.vonage.quickstart.Util.envVar;
 
-public class SendViberText {
+public class SendWhatsappSticker {
 
 	public static void main(String[] args) throws Exception {
 		configureLogging();
@@ -38,11 +37,11 @@ public class SendViberText {
 				.build()
 				.getMessagesClient()
 				.useSandboxEndpoint()
-				.sendMessage(ViberTextRequest.builder()
-						.from(envVar("FROM_ID")).to(envVar("TO_NUMBER"))
-						.text("Don't miss out on our latest offers!")
-						.category(Category.PROMOTION)
-						.build()
+				.sendMessage(WhatsappStickerRequest.builder()
+					.from(envVar("MESSAGES_SANDBOX_WHATSAPP_NUMBER"))
+					.to(envVar("MESSAGES_SANDBOX_ALLOW_LISTED_TO_NUMBER"))
+					.url("https://file-examples.com/storage/fe0b804ac5640668798b8d0/2020/03/file_example_WEBP_250kB.webp")
+					.build()
 				).getMessageUuid()
 		);
 	}
